# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  rest.ErrResponse:
    properties:
      error:
        $ref: '#/definitions/rest.Err'
    type: object
...
  rest.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  rest.TaskCreateRequest:
    properties:
      description:
        type: string
      due_date:
        type: string
      estimated_minutes:
        type: integer
      priority:
        type: string
      title:
        type: string
    type: object
  rest.TaskDetailsResponse:
    properties:
      tags:
        items:
          $ref: '#/definitions/rest.TaskAssignedTagResponse'
        type: array
      task:
        $ref: '#/definitions/rest.TaskResponse'
    type: object
  rest.TaskListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/rest.TaskResponse'
        type: array
      next_page_token:
        type: string
    type: object
  rest.TaskResponse:
    properties:
      actual_minutes:
        type: integer
      assignee_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      estimated_minutes:
        type: integer
      id:
        type: string
      priority:
        type: string
      progress:
        type: integer
      project_id:
        type: string
      revision:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  rest.TaskUpdateRequest:
    properties:
      actual_minutes:
        $ref: '#/definitions/rest.optionalJSON-int'
      description:
        $ref: '#/definitions/rest.optionalJSON-string'
      due_date:
        $ref: '#/definitions/rest.optionalJSON-string'
      estimated_minutes:
        $ref: '#/definitions/rest.optionalJSON-int'
      title:
        $ref: '#/definitions/rest.optionalJSON-string'
    type: object
...
info:
  contact: {}
  description: Task scheduling API server.
  title: task2todaytodo API
  version: "1.0"
paths:
...
  /tasks:
    get:
      description: Returns a page of tasks assigned to the authenticated user, including
        tasks in projects.
      parameters:
      - description: Items per page (default 50, maximum 100)
        in: query
        name: page_size
        type: integer
      - description: Opaque next page token
        in: query
        name: page_token
        type: string
      - description: Response field mask
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.TaskListResponse'
        "400":
          description: Invalid pagination
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to list tasks
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: List tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Creates a task without project membership.
      parameters:
      - description: Task create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.TaskCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rest.TaskResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Task ID conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to create task
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      description: Logically deletes a task the caller may delete and its child items
        and schedules.
      parameters:
      - description: Current task ETag, for example \
        in: header
        name: If-Match
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Revision conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "428":
          description: If-Match is required
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to delete task
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - Tasks
    get:
      description: Returns a task the caller may read and its assigned tags.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.TaskDetailsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to get task
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Get task
      tags:
      - Tasks
    patch:
      consumes:
      - application/json
      description: Partially updates basic task fields. Project membership, status,
        and progress cannot be changed here.
      parameters:
      - description: Current task ETag, for example \
        in: header
        name: If-Match
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.TaskUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.TaskResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Task conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "428":
          description: If-Match is required
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to update task
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - Tasks
...
  /tasks/{id}:complete:
    post:
      description: Marks task done and stops future recurrence generation.
      parameters:
      - description: Current task ETag, for example \
        in: header
        name: If-Match
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Revision conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "428":
          description: If-Match is required
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to complete task
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Complete task
      tags:
      - Tasks
  /tasks/{id}:hold:
    post:
      description: Moves task to pending, regardless of current status.
      parameters:
      - description: Current task ETag, for example \
        in: header
        name: If-Match
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Revision conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "428":
          description: If-Match is required
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to hold task
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Hold task
      tags:
      - Tasks
  /tasks/{id}:reopen:
    post:
      description: Moves task to open, regardless of current status.
      parameters:
      - description: Current task ETag, for example \
        in: header
        name: If-Match
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Revision conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "428":
          description: If-Match is required
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to reopen task
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Reopen task
      tags:
      - Tasks
  /tasks/{id}:start:
    post:
      description: Moves task to in progress, regardless of current status.
      parameters:
      - description: Current task ETag, for example \
        in: header
        name: If-Match
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Revision conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "428":
          description: If-Match is required
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to start task
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Start task
      tags:
      - Tasks
  /tasks/{id}:wait:
    post:
      description: Moves task to waiting on others, regardless of current status.
      parameters:
      - description: Current task ETag, for example \
        in: header
        name: If-Match
        required: true
        type: string
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "409":
          description: Revision conflict
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "428":
          description: If-Match is required
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Failed to wait on task
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Wait on task
      tags:
      - Tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"